refactor (Endpoints) move Endpoint/EndpointUtil out of endpoints dir
