+++ lib/Endpoint.js
@@ -1,6 +1,6 @@
-const ApiRequest = require('../ApiRequest');
+const ApiRequest = require('./ApiRequest');
 
-const ParameterError = require('../errors/ParameterError');
+const ParameterError = require('./errors/ParameterError');
 const EndpointUtil = require('./EndpointUtil');
 /**
  * Riot API Endpoint
@@ -27,9 +27,15 @@ class Endpoint {
 		this.apiVersion = this.config.API_VERSION;
 		// NOTE: will be amended by implementing Endpoints
 		this.apiUrl = `https://{platformId}.${this.config.API_HOST}/lol`;
+		this.requestOptions = {token: this.config.API_KEY};
+		if (this.config.limits.retryEndpoints.indexOf(this.name) !== -1){
+			this.requestOptions.numRetriesLeft = this.config.limits.numMaxRetries;
+			this.requestOptions.intervalRetryMS = this.config.limits.intervalRetryMS;
+		}
 
 		if (this.config.caching.isEnabled) {this.setCache(this.config.caching.defaults);}
 		if (rateLimiter) { this.rateLimiter = rateLimiter; }
+
 	}
 
 	enableCaching(options = this.config.caching.defaults, Constructor = this.config.caching.constructor) {
@@ -101,7 +107,7 @@ class Endpoint {
 		} else {
 			return this.rateLimiter.per10[platformId].scheduling(
 				() => this.rateLimiter.per600[platformId].scheduling(
-					() => ApiRequest.executing(requestUrl, {token: this.config.API_KEY})
+					() => ApiRequest.executing(requestUrl, this.requestOptions)
 				)
 			).then(response => {
 				if (this.cache) {this.cache.set(requestUrl, response);}
@@ -129,7 +135,7 @@ class Endpoint {
 			return Promise.resolve(cacheValue);
 		} else {
 
-			return ApiRequest.executing(requestUrl, {token: this.config.API_KEY})
+			return ApiRequest.executing(requestUrl, this.requestOptions)
 				.then(response => {
 					if (this.cache) {this.cache.set(requestUrl, response);}
 					return response;
