+++ lib/EndpointUtil.js
@@ -1,6 +1,6 @@
-const ParameterError = require('../errors/ParameterError');
-const LeagueUtil = require('../LeagueUtil');
-const RateLimiter = require('../RateLimiter');
+const ParameterError = require('./errors/ParameterError');
+const LeagueUtil = require('./LeagueUtil');
+const RateLimiter = require('./RateLimiter');
 
 class EndpointUtil {
