+++ lib/endpoints/ChampionEndpoint.js
@@ -1,5 +1,5 @@
-const Endpoint = require ('./Endpoint');
-const EndpointUtil = require('./EndpointUtil');
+const Endpoint = require ('../Endpoint');
+const EndpointUtil = require('../EndpointUtil');
 
 /***
  * Endpoint to receive client information about champions on the league platform
