+++ lib/endpoints/ChampionMasteryEndpoint.js
@@ -1,5 +1,5 @@
-const Endpoint = require ('./Endpoint');
-const EndpointUtil = require('./EndpointUtil');
+const Endpoint = require ('../Endpoint');
+const EndpointUtil = require('../EndpointUtil');
 
 /**
  * Endpoint to receive ChampionMastery information
