+++ lib/endpoints/LeagueEndpoint.js
@@ -1,5 +1,5 @@
-const Endpoint = require ('./Endpoint');
-const EndpointUtil = require('./EndpointUtil');
+const Endpoint = require ('../Endpoint');
+const EndpointUtil = require('../EndpointUtil');
 
 const ParameterError = require('../errors/ParameterError');
